// src/@types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IArticleFields {
  /** title */
  title: string;

  /** slug */
  slug: string;

  /** Body */
  body: Document;

  /** Image */
  image: {
    fields: {
      file: {
        url: string;
      };
    };
  };

  /** date */
  date: string;

  /** author */
  author: IAuthor;
}

export interface IArticle extends Entry<IArticleFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "article";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
  fields: IArticleFields;
}

export interface IAuthorFields {
  /** Name */
  name: string;

  /** Image */
  image: {
    fields: {
      file: {
        url: string;
      };
    };
  };
}

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "author";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
  fields: IAuthorFields;
}

export type CONTENT_TYPE = "article" | "author";

export type IEntry = IArticle | IAuthor;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";